[Feature] Add artificial light based on the daylight hours.
So in the configuraiton tab the gardener can set the desired hours
in which light above threshold is required for the plant.

If there is enough natural light for the desired hours no artifical light
switched on. Just in case the desired hours with minimum threshold lux is not
reached the difference is added via the artifical light.

On Midnight the daylight counter is logged and resetted.